feat(close-long)!: add the ability to partially close a position (#130)
* refactor(positions): replace the leverage by the position expo in position and action structs

BREAKING CHANGE: Position and PendingAction structs do not return the leverage anymore, they have
the position expo instead

* chore: Add todo and remove console.log import

* test: Move _positionValue fuzz test to its own file

* test: add unit tests to _calculatePositionExpo

* test: fix some comments and unused imports

* refactor(position): replace expo with totalexpo toavoid confusions in position structs

* refactor: revert when liquidation price is lower than start price in position expo calculation

* test(positionvalue): refactor fuzzing test for positionvalue()

* test(positionvalue): add fuzzing test comparing current and former implementation of _positionvalue

* docs: fix some comments

* fix: recalculate the leverage before firing the ValidatedOpenPosition event

* docs(natspec): improve NatSpecs

* refactor(longpendingaction): rename totalExpo to closetotalexpo for consistency

* docs(longpendingaction): re-order the params in the NatSpec to be consistent with the type

* feat(long): allow to partially close a position

initiateClosePosition will now take an additional argument that allows the user to choose an amount
to subtract from his position's size, instead of closing the entire position.

BREAKING CHANGE: initiateClosePosition takes a new argument `uint128 amountToClose`. The
InitiatedClosePosition event now also provides the amount and total expo that was removed from the
position.

* test(partial-close): add unit tests for the partial close of a position

* refactor(remove-position): fuse the _removePosition and _removeAmountFromPosition functions

* refactor(close-position): the InitiatedClosePosition now give the remaining amount and total expo

BREAKING CHANGE: More data in the InitiatedClosePosition event

* docs: add some details in the events and functions docs

* test: use the new options in the params to disable fees and fundings in tests

* chore: remove unused imports in new tests

* chore: add an Actions subdirectory to the USDNProtocol tests

* chore: some adjustments in comments and test scenarios

* test: replace calls to validateClosePosition so _validateClosePosition

* chore: use pascal case for test file names and move Open.t.sol to the Actions subdirectory

* test(partial-close): add a test for the prtial close of a position in profit after the initiate

* docs: fix an incorrect NatSpec

* chore: Rename totalExpoToClose to totalExpoToRemove for consistency

* docs: fix and add some comments

* chore: revert some unnecessary changes

* test(partial-close): add some checks on the amount of total long positions

* chore: fix typo

* refactor(close-position): use helper to setup long positions in the tests

* test(partial-close): remove duplicate check on user's balance in test

* chore: fix typo and wrong type for position size in setup long position helper

* refactor: safe cast the result of fullMulDiv

* test: use params variable instead of DEFAULT_PARAMS constant in tests

* test(partial-close): add test case for position with losses

* test: use params instead of DEFAULT_PARAMS in close position fuzzing tests

* docs: add missing NatSpec and fix typo

* chore: rename some variables and use setup helpers in tests

* test(partial-close): add check on total expo by tick

* refactor: Add a parameter to _removeAmountFromPosition to avoid double calculation of totalexpo

* refactor(partial-position): remove additional positions to compensate the double subtraction (#138)

* refactor: inverse arguments of the UsdnProtocolAmountToCloseHigherThanPositionAmount error

To be more consistemt with the name (amountToClose before position amount)

* refactor: various optimizations and fixes

* fix: fix test missing an argument in the initiateClosePosition function

* test: fix tests after merging main

* docs(close-position): specify that the amount to close is based on the amount of collateral in the

position

* refactor: add an unchecked to the position amount subtraction

* refactor(close-position): add an unchecked block to the decrementation of long position counters

+++ test/unit/UsdnProtocol/Actions/Open.t.sol
@@ -66,7 +66,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
             0
         ); // expected event
         (int24 tick, uint256 tickVersion, uint256 index) =
-            protocol.initiateOpenPosition(uint96(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
+            protocol.initiateOpenPosition(uint128(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
         uint256 tickLiqPrice = protocol.getEffectivePriceForTick(
             tick - int24(protocol.getLiquidationPenalty()) * protocol.getTickSpacing()
         );
@@ -123,7 +123,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
      */
     function test_RevertWhen_initiateOpenPositionLowLeverage() public {
         vm.expectRevert(UsdnProtocolLeverageTooLow.selector);
-        protocol.initiateOpenPosition(uint96(LONG_AMOUNT), 100_000, abi.encode(CURRENT_PRICE), "");
+        protocol.initiateOpenPosition(uint128(LONG_AMOUNT), 100_000, abi.encode(CURRENT_PRICE), "");
     }
 
     /**
@@ -139,7 +139,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
         uint128 desiredLiqPrice = uint128(maxLiquidationPrice * 1.03 ether / 1 ether);
 
         vm.expectRevert(UsdnProtocolLeverageTooHigh.selector);
-        protocol.initiateOpenPosition(uint96(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
+        protocol.initiateOpenPosition(uint128(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
     }
 
     /**
@@ -167,7 +167,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
                 UsdnProtocolLiquidationPriceSafetyMargin.selector, expectedLiqPrice, expectedMaxLiqPrice
             )
         );
-        protocol.initiateOpenPosition(uint96(LONG_AMOUNT), CURRENT_PRICE, abi.encode(CURRENT_PRICE), "");
+        protocol.initiateOpenPosition(uint128(LONG_AMOUNT), CURRENT_PRICE, abi.encode(CURRENT_PRICE), "");
     }
 
     /**
@@ -184,7 +184,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
         uint256 initialTotalExpo = protocol.getTotalExpo();
         uint128 desiredLiqPrice = CURRENT_PRICE * 2 / 3; // leverage approx 3x
         (int24 tick, uint256 tickVersion, uint256 index) =
-            protocol.initiateOpenPosition(uint96(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
+            protocol.initiateOpenPosition(uint128(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
         Position memory tempPos = protocol.getLongPosition(tick, tickVersion, index);
 
         _waitDelay();
@@ -219,7 +219,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
     function test_validateOpenPositionAboveMaxLeverage() public {
         uint128 desiredLiqPrice = CURRENT_PRICE * 9 / 10; // leverage approx 10x
         (int24 tick, uint256 tickVersion, uint256 index) =
-            protocol.initiateOpenPosition(uint96(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
+            protocol.initiateOpenPosition(uint128(LONG_AMOUNT), desiredLiqPrice, abi.encode(CURRENT_PRICE), "");
         Position memory tempPos = protocol.getLongPosition(tick, tickVersion, index);
 
         _waitDelay();
@@ -293,7 +293,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
         uint256 balanceBefore = address(this).balance;
         bytes memory priceData = abi.encode(uint128(2000 ether));
         uint256 validationCost = oracleMiddleware.validationCost(priceData, ProtocolAction.InitiateOpenPosition);
-        protocol.initiateOpenPosition{ value: 0.5 ether }(uint96(LONG_AMOUNT), 1000 ether, priceData, "");
+        protocol.initiateOpenPosition{ value: 0.5 ether }(uint128(LONG_AMOUNT), 1000 ether, priceData, "");
         assertEq(address(this).balance, balanceBefore - validationCost, "user balance after refund");
     }
 
@@ -310,7 +310,7 @@ contract TestUsdnProtocolOpenPosition is UsdnProtocolBaseFixture {
         uint128 desiredLiqPrice = CURRENT_PRICE * 2 / 3; // leverage approx 3x
         protocol.initiateOpenPosition{
             value: oracleMiddleware.validationCost(priceData, ProtocolAction.InitiateOpenPosition)
-        }(uint96(LONG_AMOUNT), desiredLiqPrice, priceData, "");
+        }(uint128(LONG_AMOUNT), desiredLiqPrice, priceData, "");
         _waitDelay();
 
         uint256 balanceBefore = address(this).balance;
